+++ src/pages/ToPage/ToPage.jsx
@@ -5,7 +5,7 @@ import Button from "../../components/common/Button/Button";
 import theme from "../../styles/theme";
 import SelectableBox from "../../components/SelectableBox/SelectableBox";
 import Input from "../../components/common/Input/Input";
-import { getBackgroundImages } from "../../api/toPageData";
+import { getBackgroundImages } from "../../api/images.api";
 import { submitToPage } from "../../api/recipient.api";
 import { showToast } from "../../components/common/Toast/Toast";
